requirements.txt
python-telegram-bot
instagrapi
requests
